+++ Vector.h
@@ -9,22 +9,22 @@ class Vector
 {
     public:
         Vector();
-        Vector(float, float, float);
+        Vector(double, double, double);
         ~Vector();
         void show();
         Vector sum(Vector);
         Vector difference(Vector);
-        float length();
-        float dotProduct(Vector);
+        double length();
+        double dotProduct(Vector);
         Vector crossProduct(Vector);
-        Vector numberProduct(float);
-        Vector numberDivision(float);
-        float getX();
-        float getY();
-        float getZ();
+        Vector numberProduct(double);
+        Vector numberDivision(double);
+        double getX();
+        double getY();
+        double getZ();
 
     private:
-        float x, y, z;
+        double x, y, z;
 };
 
 #endif // VECTOR_H
